+++ packages/headless/doc-parser/src/entity-builder.ts
@@ -13,6 +13,12 @@ interface EntityOptions extends Comment {
   isOptional: boolean;
 }
 
+interface ObjEntityOptions {
+  entity: Entity;
+  members: AnyEntity[];
+  typeName: string;
+}
+
 interface FuncEntityOptions extends Comment {
   name: string;
   params: AnyEntity[];
@@ -28,6 +34,18 @@ export function buildEntity(config: EntityOptions): Entity {
     isOptional: config.isOptional,
     type,
     desc,
+  };
+}
+
+export function buildObjEntity(config: ObjEntityOptions): ObjEntity {
+  const {entity, members} = config;
+  const typeName = sanitizeType(config.typeName);
+
+  return {
+    ...entity,
+    members,
+    isTypeExtracted: false,
+    typeName,
   };
 }