+++ packages/samples/headless-react/src/App.tsx
@@ -26,6 +26,8 @@ import {QuerySummary} from './components/query-summary/query-summary.class';
 import {QuerySummary as QuerySummaryFn} from './components/query-summary/query-summary.fn';
 import {Facet} from './components/facet/facet.class';
 import {Facet as FacetFn} from './components/facet/facet.fn';
+import {History} from './components/history/history.class';
+import {History as HistoryFn} from './components/history/history.fn';
 import {
   buildRecommendationList,
   buildTab,
@@ -43,8 +45,10 @@ import {
   SortCriterion,
   buildResultsPerPage,
   buildPager,
+  buildHistory,
 } from '@coveo/headless';
 import {bindSearchParametersToURI} from './components/search-parameter-manager/search-parameter-manager';
+import {setContext} from './components/context/context';
 
 const recommendationList = buildRecommendationList(recommendationEngine);
 
@@ -95,10 +99,13 @@ const resultsPerPage = buildResultsPerPage(engine, {
 
 const pager = buildPager(engine, {options: {numberOfPages: 6}});
 
+const history = buildHistory(engine);
+
 const {autoUpdateURI: startUpdatingURI} = bindSearchParametersToURI(engine);
 
 function App() {
   useEffect(() => startUpdatingURI(), []);
+  setContext('30-45', ['sports', 'camping', 'electronics']);
 
   return (
     <div className="App">
@@ -160,6 +167,10 @@ function App() {
         <Section title="pager">
           <Pager />
           <PagerFn controller={pager} />
+        </Section>
+        <Section title="history">
+          <History />
+          <HistoryFn controller={history} />
         </Section>
       </header>
     </div>